data buffer is now inline in struct array, removing an indirection
This also means that an array must always live on the heap since it
is a structure whose true size can not be known at compile time,
so we can dispense with the distinction between alloc and open
as well as the distinction between close and free. Only alloca and
free are required.